fix: 六角/react課程_1010-修正
